initial commit
now the version 0.0.1 is finished.
it can down from....a html file.
and the mechanism of this program is really esay,
it find all <a> element and download which has a id named "thumbnail"
.......
easy, but useful for me.
it can download whole channel page,
if you can scroll to the end of the page.